+++ sha256_mb/sha256_ctx_avx2.c
@@ -233,7 +233,7 @@ static inline uint32_t hash_pad(uint8_t padblock[SHA256_BLOCK_SIZE * 2], uint32_
 	*((uint64_t *) & padblock[i - 16]) = 0;
 #endif
 
-	*((uint64_t *) & padblock[i - 8]) = _byteswap_uint64(total_len << 3);
+	*((uint64_t *) & padblock[i - 8]) = _byteswap_uint64((uint64_t) total_len << 3);
 
 	return i >> SHA256_LOG2_BLOCK_SIZE;	// Number of extra blocks to hash
 }
